User-friendly Interface The project interface was created so simple so players can solve the puzzle without visual problem. Very Difficult Puzzle This game was meant to players that has excellence talent when solving mathematical problems. Full Mouse and Keyboard Controls The project has an easy control players can user the mouse left click for highlighting a tile and keyboard for entering the number.
Simple Sudoku Game The Simple Sudoku Game with Source Code is a single-player game where the player must solve the given blank tile by entering a combinatorial number-placement. The project purpose to enhance the mathematical capability and to test a person critical thinking. The Simple Sudoku Game was created using tkinter module whereas it display a quality interface and allow user to interact with the application. This simple game is played in a 9x9 grid board tile. The player must guess a correct number combination in order to solve the given puzzle